updated folder staructure for layout
