+++ src/app/component/Layout/Navbar/Navbar.module.css
@@ -1,23 +1,170 @@
-
-.container{
-    padding: 30px;
-    position: relative;
-    width: 78.5%;
+.header {
+    background-color: #fcf8f4;
+    padding: 16px;
+    border: 2px solid #EFDCBA;
+    width: 100vw;
+    overflow-x: hidden;
+    position: fixed;
+    top: 0;
+    left: 0;
+    z-index: 1000;
+    display: flex;
     align-items: center;
-    margin-left: 15.6rem;
-    overflow: hidden;
-    bottom: 20px;
-    background-color: #faf8f2;
-    border-bottom: 2px solid #EFDCBA;
-    /* border: 2px solid red; */
+    justify-content: space-between;
+    flex-wrap: wrap;
+}
+
+.header1 {
+    background-color: #fcf8f4;
+    padding: 16px;
+    border: 2px solid #EFDCBA;
+    width: 100vw;
+    overflow-x: hidden;
+    position: fixed;
+    top: 0;
+    left: 0;
+    z-index: 1000;
 }
 
 
+.logoContainer {
+    display: flex;
+    align-items: center;
+    flex: 1 1 auto;
+}
 
+.logoContainer1 {
+    font-size: 1.5rem;
+    font-weight: bold;
+    display: flex;
+    justify-content: flex-start;
+    align-items: center
+}
 
+.logo {
+    width: 11rem;
+}
 
+.searchBox {
+    flex-grow: 1;
+    position: relative;
+    left: 20rem;
+}
 
+.searchIcon {
+    color: #666666;
+    position: relative;
+    top: 1rem;
+    right: 2.5rem;
+    cursor: pointer;
+}
 
+.notifications {
+    display: flex;
+    align-items: center;
+    gap: 20px;
+    margin-right: 3rem;
+}
 
+.profile {
+    display: flex;
+    align-items: center;
+    gap: 10px;
+    margin-left: 1rem;
+    background-color: #784800;
+    padding: 4px;
+    border-radius: 30px;
+}
 
+.profileName {
+    color: #FFFFFF;
+    margin-left: 16px;
+}
 
+.profileBalance {
+    color: #FFFFFF;
+    font-size: 14px;
+    margin-left: 18px;
+}
+
+/* Responsiveness */
+@media (max-width: 1024px) {
+    .header {
+        padding: 15px;
+    }
+
+    .logo {
+        width: 9rem;
+    }
+
+    .profileName,
+    .profileBalance {
+        font-size: 12px;
+    }
+}
+
+@media (max-width: 768px) {
+    .header {
+        padding: 10px;
+        flex-direction: column;
+        align-items: flex-start;
+    }
+
+    .logo-container {
+        margin-bottom: 1rem;
+    }
+
+    .search-box {
+        width: 100%;
+        margin-top: 1rem;
+    }
+
+    .notifications-profile-container {
+        flex-direction: column;
+        align-items: flex-start;
+        margin-top: 1rem;
+    }
+
+    .profile-container {
+        width: 100%;
+        justify-content: flex-start;
+    }
+
+    .search-box input {
+        padding: 8px 12px;
+    }
+}
+
+@media (max-width: 480px) {
+    .header {
+        padding: 8px;
+    }
+
+    .logo {
+        width: 8rem;
+    }
+
+    .search-box {
+        margin-top: 0.5rem;
+    }
+
+    .notifications-profile-container {
+        gap: 10px;
+        flex-direction: column;
+        align-items: flex-start;
+    }
+
+    .profile-name,
+    .profile-balance {
+        font-size: 12px;
+    }
+
+    .search-box input {
+        font-size: 14px;
+        padding: 8px;
+    }
+
+    .search-icon {
+        right: 5px;
+    }
+}
