+++ src/app/component/layout.js
@@ -22,7 +22,7 @@ const Layout = ({ children }) => {
             <Sidebar />
           </div>
           <div className={styles.content}>
-            {/* <Navbar /> */}
+             <Navbar />
             {children}
             {/* <Footer /> */}
           </div>
