+++ src/app/login/Navbar/Navbar.module.css
@@ -8,32 +8,8 @@
     top: 0;
     left: 0;
     z-index: 1000;
-    display: flex;
-    align-items: center;
-    justify-content: space-between;
-    flex-wrap: wrap;
 }
-
-.header1 {
-    background-color: #fcf8f4;
-    padding: 16px;
-    border: 2px solid #EFDCBA;
-    width: 100vw;
-    overflow-x: hidden;
-    position: fixed;
-    top: 0;
-    left: 0;
-    z-index: 1000;
-}
-
-
 .logoContainer {
-    display: flex;
-    align-items: center;
-    flex: 1 1 auto;
-}
-
-.logoContainer1 {
     font-size: 1.5rem;
     font-weight: bold;
     display: flex;
@@ -43,48 +19,6 @@
 
 .logo {
     width: 11rem;
-}
-
-.searchBox {
-    flex-grow: 1;
-    position: relative;
-    left: 20rem;
-}
-
-.searchIcon {
-    color: #666666;
-    position: relative;
-    top: 1rem;
-    right: 2.5rem;
-    cursor: pointer;
-}
-
-.notifications {
-    display: flex;
-    align-items: center;
-    gap: 20px;
-    margin-right: 3rem;
-}
-
-.profile {
-    display: flex;
-    align-items: center;
-    gap: 10px;
-    margin-left: 1rem;
-    background-color: #784800;
-    padding: 4px;
-    border-radius: 30px;
-}
-
-.profileName {
-    color: #FFFFFF;
-    margin-left: 16px;
-}
-
-.profileBalance {
-    color: #FFFFFF;
-    font-size: 14px;
-    margin-left: 18px;
 }
 
 /* Responsiveness */
